chore: added common preview component
